requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
filelock==3.15.3
gunicorn==22.0.0
identify==2.5.36
nodeenv==1.9.1
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
virtualenv==20.26.2
whitenoise==6.6.0
